normalisasi
