--- composer.json.dist
@@ -1,11 +1,0 @@
-{
-  "description": "CodeIgniter Xtend Starter",
-  "require": {
-    "php": "^7.0",
-    "buddywinangun/codeigniter-xtend": "@dev"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "^9.6",
-    "codedungeon/phpunit-result-printer": "^0.32.0"
-  }
-}
