--- src/RESTful/Exception/RestClient.php
@@ -1,14 +1,0 @@
-<?php
-
-/**
- * This file is part of Codeigniter Xtend.
- *
- * For the full copyright and license information, please view
- * the LICENSE file that was distributed with this source code.
- */
-
-namespace Xtend\RESTful\Exception;
-
-class RestClient extends \Exception
-{
-}
