--- start/application/config/locate.php
@@ -1,18 +1,0 @@
-<?php
-defined('BASEPATH') or exit('No direct script access allowed');
-
-/*
-| -------------------------------------------------------------------
-|  Packages
-| -------------------------------------------------------------------
-| This maps the locations of any packages in your application
-*/
-$locate['packages'] = array();
-
-/*
-| -------------------------------------------------------------------
-|  Template
-| -------------------------------------------------------------------
-| This maps the locations of any template in your application
-*/
-$locate['template'] = array();
