--- start/application/core/AppConfig.php
@@ -1,6 +1,0 @@
-<?php
-defined('BASEPATH') or exit('No direct script access allowed');
-
-class AppConfig extends \Xtend\Config\Config
-{
-}
