--- start/application/core/AppController.php
@@ -1,7 +1,0 @@
-<?php
-defined('BASEPATH') or exit('No direct script access allowed');
-
-abstract class AppController extends \Xtend\Controller
-{
-  protected function beforeResponse(string $referer) {}
-}
