--- start/application/core/AppHooks.php
@@ -1,6 +1,0 @@
-<?php
-defined('BASEPATH') or exit('No direct script access allowed');
-
-class AppHooks extends \Xtend\Events\Hooks
-{
-}
