--- start/application/core/AppLang.php
@@ -1,6 +1,0 @@
-<?php
-defined('BASEPATH') or exit('No direct script access allowed');
-
-class AppLang extends Xtend\Language\Lang
-{
-}
