--- start/application/core/AppLoader.php
@@ -1,6 +1,0 @@
-<?php
-defined('BASEPATH') or exit('No direct script access allowed');
-
-class AppLoader extends \Xtend\Loader
-{
-}
