--- start/application/core/AppModel.php
@@ -1,6 +1,0 @@
-<?php
-defined('BASEPATH') or exit('No direct script access allowed');
-
-abstract class AppModel extends \Xtend\Model
-{
-}
