--- start/application/core/AppRouter.php
@@ -1,6 +1,0 @@
-<?php
-defined('BASEPATH') or exit('No direct script access allowed');
-
-class AppRouter extends \Xtend\Router\Router
-{
-}
