--- start/application/language/english/english_lang.php
@@ -1,14 +1,0 @@
-<?php
-
-defined('BASEPATH') or exit('No direct script access allowed');
-
-/*
-Name: English
-name_en: English
-locale: en-US
-currency: €
-currency_key: EUR
-direction: ltr
-code: en
-flag: us
-*/
