--- start/application/libraries/AppForm_validation.php
@@ -1,8 +1,0 @@
-<?php
-defined('BASEPATH') or exit('No direct script access allowed');
-
-use Xtend\Validation\FormValidation;
-
-class AppForm_validation extends FormValidation
-{
-}
